Fix clippy issues

+++ blst/src/kzg_proofs.rs
@@ -58,7 +58,7 @@ pub fn g1_linear_combination(out: &mut FsG1, points: &[FsG1], scalars: &[FsFr], 
     {
         let mut scratch: Vec<u8>;
         unsafe {
-            scratch = vec![0u8; blst_p1s_mult_pippenger_scratch_sizeof(len) as usize];
+            scratch = vec![0u8; blst_p1s_mult_pippenger_scratch_sizeof(len)];
         }
 
         let mut p_affine = vec![blst_p1_affine::default(); len];
